# actions/openapi.github-contents.yaml
openapi: 3.1.0
info:
  title: GitHub Contents API (GET/PUT)
  version: 1.0.0
servers:
  - url: https://api.github.com
paths:
  /repos/{owner}/{repo}/contents/{path}:
    get:
      operationId: getRepoFile
      summary: Get a file from a GitHub repo (Contents API)
      parameters:
        - in: path
          name: owner
          required: true
          schema: { type: string }
        - in: path
          name: repo
          required: true
          schema: { type: string }
        - in: path
          name: path
          required: true
          schema: { type: string }
        - in: query
          name: ref
          required: false
          schema: { type: string }
      responses:
        "200":
          description: File content + sha
          content:
            application/json:
              schema:
                type: object
                properties:
                  sha: { type: string }
                  content: { type: string }
                  encoding: { type: string }
    put:
      operationId: putRepoFile
      summary: Create or update a file in a GitHub repo (Contents API)
      parameters:
        - in: path
          name: owner
          required: true
          schema: { type: string }
        - in: path
          name: repo
          required: true
          schema: { type: string }
        - in: path
          name: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, content]
              properties:
                message: { type: string }
                content:
                  type: string
                  description: Base64-encoded file content
                sha:
                  type: string
                  description: Required for updates (from GET)
                branch:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: object
                    properties:
                      sha: { type: string }
                  commit:
                    type: object
                    properties:
                      sha: { type: string }
components:
  securitySchemes:
    githubToken:
      type: http
      scheme: bearer
      bearerFormat: token
security:
  - githubToken: []